+++ oximeter/types/versions/src/impls/mod.rs
@@ -4,8 +4,12 @@
 
 //! Functional code for the latest versions of types.
 
+pub(crate) mod histogram;
 mod producer;
-pub mod traits;
+pub(crate) mod quantile;
+pub(crate) mod schema;
+pub(crate) mod traits;
+pub(crate) mod types;
 
 use crate::latest::schema::TimeseriesName;
 use crate::latest::types::MetricsError;
