+++ oximeter/types/versions/src/initial/schema.rs
@@ -6,10 +6,6 @@
 
 use super::types::DatumType;
 use super::types::FieldType;
-use super::types::MetricsError;
-use super::types::Sample;
-use crate::impls::traits::Metric;
-use crate::impls::traits::Target;
 use chrono::DateTime;
 use chrono::Utc;
 use parse_display::Display;
@@ -38,13 +34,6 @@ pub struct FieldSchema {
     pub field_type: FieldType,
     pub source: FieldSource,
     pub description: String,
-}
-
-impl FieldSchema {
-    /// Return `true` if this field is copyable.
-    pub const fn is_copyable(&self) -> bool {
-        self.field_type.is_copyable()
-    }
 }
 
 /// The source from which a field is derived, the target or metric.
@@ -110,64 +99,6 @@ impl JsonSchema for TimeseriesName {
     }
 }
 
-impl std::ops::Deref for TimeseriesName {
-    type Target = String;
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl From<TimeseriesName> for String {
-    fn from(n: TimeseriesName) -> Self {
-        n.0
-    }
-}
-
-impl std::fmt::Display for TimeseriesName {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
-
-impl std::convert::TryFrom<&str> for TimeseriesName {
-    type Error = MetricsError;
-    fn try_from(s: &str) -> Result<Self, Self::Error> {
-        validate_timeseries_name(s).map(|s| TimeseriesName(s.to_string()))
-    }
-}
-
-impl std::convert::TryFrom<String> for TimeseriesName {
-    type Error = MetricsError;
-    fn try_from(s: String) -> Result<Self, Self::Error> {
-        validate_timeseries_name(&s)?;
-        Ok(TimeseriesName(s))
-    }
-}
-
-impl std::str::FromStr for TimeseriesName {
-    type Err = MetricsError;
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        s.try_into()
-    }
-}
-
-impl<T> PartialEq<T> for TimeseriesName
-where
-    T: AsRef<str>,
-{
-    fn eq(&self, other: &T) -> bool {
-        self.0.eq(other.as_ref())
-    }
-}
-
-fn validate_timeseries_name(s: &str) -> Result<&str, MetricsError> {
-    if regex::Regex::new(TIMESERIES_NAME_REGEX).unwrap().is_match(s) {
-        Ok(s)
-    } else {
-        Err(MetricsError::InvalidTimeseriesName)
-    }
-}
-
 /// Text descriptions for the target and metric of a timeseries.
 #[derive(Clone, Debug, Default, Deserialize, JsonSchema, Serialize)]
 pub struct TimeseriesDescription {
@@ -211,148 +142,6 @@ pub struct TimeseriesSchema {
     pub created: DateTime<Utc>,
 }
 
-/// Default version for timeseries schema, 1.
-pub const fn default_schema_version() -> NonZeroU8 {
-    unsafe { NonZeroU8::new_unchecked(1) }
-}
-
-impl From<&Sample> for TimeseriesSchema {
-    fn from(sample: &Sample) -> Self {
-        let timeseries_name = sample
-            .timeseries_name
-            .parse()
-            .expect("expected a legal timeseries name in a sample");
-        let mut field_schema = BTreeSet::new();
-        for field in sample.target_fields() {
-            let schema = FieldSchema {
-                name: field.name.clone(),
-                field_type: field.value.field_type(),
-                source: FieldSource::Target,
-                description: String::new(),
-            };
-            field_schema.insert(schema);
-        }
-        for field in sample.metric_fields() {
-            let schema = FieldSchema {
-                name: field.name.clone(),
-                field_type: field.value.field_type(),
-                source: FieldSource::Metric,
-                description: String::new(),
-            };
-            field_schema.insert(schema);
-        }
-        let datum_type = sample.measurement.datum_type();
-        Self {
-            timeseries_name,
-            description: Default::default(),
-            field_schema,
-            datum_type,
-            version: default_schema_version(),
-            authz_scope: AuthzScope::Fleet,
-            units: Units::Count,
-            created: Utc::now(),
-        }
-    }
-}
-
-impl TimeseriesSchema {
-    /// Construct a timeseries schema from a target and metric.
-    pub fn new<T, M>(target: &T, metric: &M) -> Result<Self, MetricsError>
-    where
-        T: Target,
-        M: Metric,
-    {
-        let timeseries_name = crate::impls::timeseries_name(target, metric)?;
-        let mut field_schema = BTreeSet::new();
-        for field in target.fields() {
-            let schema = FieldSchema {
-                name: field.name.clone(),
-                field_type: field.value.field_type(),
-                source: FieldSource::Target,
-                description: String::new(),
-            };
-            field_schema.insert(schema);
-        }
-        for field in metric.fields() {
-            let schema = FieldSchema {
-                name: field.name.clone(),
-                field_type: field.value.field_type(),
-                source: FieldSource::Metric,
-                description: String::new(),
-            };
-            field_schema.insert(schema);
-        }
-        let datum_type = metric.datum_type();
-        Ok(Self {
-            timeseries_name,
-            description: Default::default(),
-            field_schema,
-            datum_type,
-            version: default_schema_version(),
-            authz_scope: AuthzScope::Fleet,
-            units: Units::Count,
-            created: Utc::now(),
-        })
-    }
-
-    /// Construct a timeseries schema from a sample
-    pub fn from_sample(sample: &Sample) -> Self {
-        Self::from(sample)
-    }
-
-    /// Return the schema for the given field.
-    pub fn schema_for_field<S>(&self, name: S) -> Option<&FieldSchema>
-    where
-        S: AsRef<str>,
-    {
-        self.field_schema.iter().find(|field| field.name == name.as_ref())
-    }
-
-    /// Return an iterator over the target fields.
-    pub fn target_fields(&self) -> impl Iterator<Item = &FieldSchema> {
-        self.field_iter(FieldSource::Target)
-    }
-
-    /// Return an iterator over the metric fields.
-    pub fn metric_fields(&self) -> impl Iterator<Item = &FieldSchema> {
-        self.field_iter(FieldSource::Metric)
-    }
-
-    /// Return an iterator over fields from the given source.
-    fn field_iter(
-        &self,
-        source: FieldSource,
-    ) -> impl Iterator<Item = &FieldSchema> {
-        self.field_schema.iter().filter(move |field| field.source == source)
-    }
-
-    /// Return the target and metric component names for this timeseries
-    pub fn component_names(&self) -> (&str, &str) {
-        self.timeseries_name
-            .split_once(':')
-            .expect("Incorrectly formatted timseries name")
-    }
-
-    /// Return the name of the target for this timeseries.
-    pub fn target_name(&self) -> &str {
-        self.component_names().0
-    }
-
-    /// Return the name of the metric for this timeseries.
-    pub fn metric_name(&self) -> &str {
-        self.component_names().1
-    }
-}
-
-impl PartialEq for TimeseriesSchema {
-    fn eq(&self, other: &TimeseriesSchema) -> bool {
-        self.timeseries_name == other.timeseries_name
-            && self.version == other.version
-            && self.datum_type == other.datum_type
-            && self.field_schema == other.field_schema
-    }
-}
-
 // Regular expression describing valid timeseries names.
 //
 // Names are derived from the names of the Rust structs for the target and metric, converted to
@@ -363,7 +152,7 @@ impl PartialEq for TimeseriesSchema {
 //  - Zero or more of the above, delimited by '-'.
 //
 // That describes the target/metric name, and the timeseries is two of those, joined with ':'.
-const TIMESERIES_NAME_REGEX: &str =
+pub(crate) const TIMESERIES_NAME_REGEX: &str =
     "^(([a-z]+[a-z0-9]*)(_([a-z0-9]+))*):(([a-z]+[a-z0-9]*)(_([a-z0-9]+))*)$";
 
 /// Authorization scope for a timeseries.
@@ -396,53 +185,4 @@ pub enum AuthzScope {
     Project,
     /// The timeseries is viewable to all without limitation.
     ViewableToAll,
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::convert::TryFrom;
-
-    #[test]
-    fn test_timeseries_name() {
-        let name = TimeseriesName::try_from("foo:bar").unwrap();
-        assert_eq!(format!("{}", name), "foo:bar");
-    }
-
-    #[test]
-    fn test_timeseries_name_from_str() {
-        assert!(TimeseriesName::try_from("a:b").is_ok());
-        assert!(TimeseriesName::try_from("a_a:b_b").is_ok());
-        assert!(TimeseriesName::try_from("a0:b0").is_ok());
-        assert!(TimeseriesName::try_from("a_0:b_0").is_ok());
-
-        assert!(TimeseriesName::try_from("_:b").is_err());
-        assert!(TimeseriesName::try_from("a_:b").is_err());
-        assert!(TimeseriesName::try_from("0:b").is_err());
-        assert!(TimeseriesName::try_from(":b").is_err());
-        assert!(TimeseriesName::try_from("a:").is_err());
-        assert!(TimeseriesName::try_from("123").is_err());
-        assert!(TimeseriesName::try_from("x.a:b").is_err());
-    }
-
-    #[test]
-    fn test_field_schema_ordering() {
-        let mut fields = BTreeSet::new();
-        fields.insert(FieldSchema {
-            name: String::from("second"),
-            field_type: FieldType::U64,
-            source: FieldSource::Target,
-            description: String::new(),
-        });
-        fields.insert(FieldSchema {
-            name: String::from("first"),
-            field_type: FieldType::U64,
-            source: FieldSource::Target,
-            description: String::new(),
-        });
-        let mut iter = fields.iter();
-        assert_eq!(iter.next().unwrap().name, "first");
-        assert_eq!(iter.next().unwrap().name, "second");
-        assert!(iter.next().is_none());
-    }
 }
