+++ oximeter/types/versions/src/latest.rs
@@ -11,14 +11,15 @@ pub mod collector {
 pub mod histogram {
     pub use crate::v1::histogram::Bin;
     pub use crate::v1::histogram::BinRange;
-    pub use crate::v1::histogram::Bits;
     pub use crate::v1::histogram::Histogram;
-    pub use crate::v1::histogram::HistogramAdditiveWidth;
-    pub use crate::v1::histogram::HistogramError;
-    pub use crate::v1::histogram::HistogramSupport;
-    pub use crate::v1::histogram::LogLinearBins;
-    pub use crate::v1::histogram::QuantizationError;
-    pub use crate::v1::histogram::Record;
+
+    pub use crate::impls::histogram::Bits;
+    pub use crate::impls::histogram::HistogramAdditiveWidth;
+    pub use crate::impls::histogram::HistogramError;
+    pub use crate::impls::histogram::HistogramSupport;
+    pub use crate::impls::histogram::LogLinearBins;
+    pub use crate::impls::histogram::QuantizationError;
+    pub use crate::impls::histogram::Record;
 }
 
 pub mod producer {
@@ -31,7 +32,11 @@ pub mod producer {
 
 pub mod quantile {
     pub use crate::v1::quantile::Quantile;
-    pub use crate::v1::quantile::QuantileError;
+
+    pub use crate::impls::quantile::QuantileError;
+
+    // Use by both type definition and impls, but not pub.
+    pub(crate) use crate::v1::quantile::FILLED_MARKER_LEN;
 }
 
 pub mod schema {
@@ -43,7 +48,11 @@ pub mod schema {
     pub use crate::v1::schema::TimeseriesName;
     pub use crate::v1::schema::TimeseriesSchema;
     pub use crate::v1::schema::Units;
-    pub use crate::v1::schema::default_schema_version;
+
+    pub use crate::impls::schema::default_schema_version;
+
+    // Use by both type definition and impls, but not pub.
+    pub(crate) use crate::v1::schema::TIMESERIES_NAME_REGEX;
 }
 
 pub mod traits {
@@ -63,9 +72,14 @@ pub mod types {
     pub use crate::v1::types::FieldType;
     pub use crate::v1::types::FieldValue;
     pub use crate::v1::types::Measurement;
-    pub use crate::v1::types::MetricsError;
     pub use crate::v1::types::MissingDatum;
     pub use crate::v1::types::ProducerRegistry;
+    pub use crate::v1::types::ProducerResults;
     pub use crate::v1::types::ProducerResultsItem;
     pub use crate::v1::types::Sample;
+
+    pub use crate::impls::types::MetricsError;
+
+    // Use by both type definition and impls, but not pub.
+    pub(crate) use crate::v1::types::FieldSet;
 }
